fix(3p/lisp/mime4cl): fix stat function name in FILE-SIZE
This is a silly mistake and was not caught because FILE-SIZE isn't
exercised in the test suite. We can probably remove MIME-BODY-SIZE and
look into removing MIME-PART-SIZE as well. I just want to be thorough
here so that we can revert into a non-broken state in case we decide we
need those functions for something.

Change-Id: I5bbb3dde6616220fc3b6feddbf7a39b6a9b0ea0d
Reviewed-on: https://cl.tvl.fyi/c/depot/+/12861
Autosubmit: sterni <sternenseemann@systemli.org>
Reviewed-by: sterni <sternenseemann@systemli.org>
Tested-by: BuildkiteCI