feat(sterni/emacs): project-independent project-find-regexp
I like project-find-regexp better than grep-find-file, but it's a little
annoying to use for arbitrary directories. With C-u, it unnecessarily
prompts for a file name pattern.

For convenience, add a variant of project-find-regexp which always
prompts for a directory, but doesn't require a project nor prompts for a
file name pattern.

Change-Id: I40f0bdcca4aa129f447f4f4790d732afcf76d3f7
Reviewed-on: https://cl.tvl.fyi/c/depot/+/13056
Autosubmit: sterni <sternenseemann@systemli.org>
Reviewed-by: sterni <sternenseemann@systemli.org>
Tested-by: BuildkiteCI