feat(3p/overlays): build Nix 2.3 from TVL maintenance branch
Our maintenance branch includes additional concurrency fixes which are not in
the upstream 2.3 branch.

These issues are fixed in C++ Nix HEAD, but in a more invasive way (by removing
the second set of locks completely).

This also retains additional debug information in the built binaries to make
future issues easier to debug.

Change-Id: I4e7a8baabd059c96404822d9634df52c403a869f
Reviewed-on: https://cl.tvl.fyi/c/depot/+/13135
Tested-by: BuildkiteCI
Reviewed-by: sterni <sternenseemann@systemli.org>