docs(snix): move some more docs
Absorb some of the paragraphs from snix/docs/src/architecture.md, as
well as the blog post around the builder protocol, and extend it with a
bit more context.

The parts about the Evaluator are largely outdated (see the warning
above in that file), and otherwise already covered.

Update and restructure protocol.md to describe the goals of the builder.

Change-Id: I26ed625f7b6fcfded3e65cf87b00ceb33b75f229
Reviewed-on: https://cl.snix.dev/c/snix/+/30264
Reviewed-by: edef . <edef@edef.eu>
Tested-by: besadii
Autosubmit: Florian Klink <flokli@flokli.de>