fix(snix/boot/tests): add name attr to mkBootTest
Previously when running all boot integration tests with
`mg build //snix/boot:tests`, and one of them failing, it was not clear
which one failed, because on failed test derivation name shown used the
same suffix for all tests. Now, for a cost of a bit more boilerplate,
test derivation name contains its name, so it's immediately obvious.

Change-Id: If8766e12048a50f21d61c615239ee62b3361ac2f
Reviewed-on: https://cl.snix.dev/c/snix/+/30674
Reviewed-by: Florian Klink <flokli@flokli.de>
Tested-by: besadii