feat(verify-lang-tests): add glue
So far we apparently didn't automatically cross-check our test files
in snix-glue against Nix 2.3 and Nix 2.30 🙈.

Instead of setting testRoot at the top, pass it around as an additional
argument, and instantiate it for glue too.

While at it, change the attrset keys as well to remove the version interpolation.
It looks like we plan to always test against 2.3 and the latest
(verified) version, so we can just make this static, and not use a dot,
which made this attrset hard to call from the CLI.

As the result of these changes, update glue cases so that they match
what CppNix returns.

Add eval-okay-parseFlakeRef.flags to make verify-lang-tests invoke nix
with the flakes experimental feature, so it's able to use builtins.parseFlakeRef

Also, remove the NIX_STORE_DIR env in the sandbox. It does not seem to
break anything, but enables some tests e.g.
eval-okay-toxml-*. This kind of tests require store path
assertion and expect the store directory to be /nix/store. But with different
store directory, derivations produce competely different output hashes,
so we can't just replace certain output parts (/nix/store with randomly generated
NIX_STORE_DIR in that case) before comparison as we do with some other
tests.

Fixes https://git.snix.dev/snix/snix/issues/171

Co-Authored-By: Vladimir Fetisov <vladimir.fetisov@pm.me>
Change-Id: Ic46d0ed5881cd87493a90273b94de98481c1d207
Reviewed-on: https://cl.snix.dev/c/snix/+/30700
Tested-by: besadii
Reviewed-by: Florian Klink <flokli@flokli.de>
Autosubmit: Florian Klink <flokli@flokli.de>