fix(eval/cli): do not cast to string upon `--raw`
With this change
```sh
cargo run --bin snix -- --expr '2' --raw
```
no longer panics with
```
called `Result::unwrap()` on an `Err` value: TypeError { expected: "contextful string", actual: "int" }
```

This extends to any other Nix expression that does not evaluate to a
string.

This is because `to_contextful_str` expects a `NixString`.

Change-Id: I6a6a696449b889bae272abafd0d5227e6efe2e66
Reviewed-on: https://cl.snix.dev/c/snix/+/30726
Reviewed-by: Florian Klink <flokli@flokli.de>
Tested-by: besadii